mongo java demo added

+++ rest-server-client2-microunits/departmentmgt-server/src/main/java/com/dxctraining/deptmgt/department/entities/Department.java
@@ -41,6 +41,6 @@ public class Department {
 
     @Override
     public int hashCode() {
-        return id.intValue();
+        return id;
     }
 }
